LU-10560 llite: remove extra headers from llite_mmap.c
Remove headers from llite_mmap.c that are no longer (or were possibly
never) needed.  This avoids a compile problem with 4.14 kernels:

	CC [M]  lustre/llite/llite_mmap.o
	In file included from lustre/llite/llite_mmap.c:41:0:
	./arch/x86/include/asm/uaccess.h: In function ‘set_fs’:
	./arch/x86/include/asm/uaccess.h:31:9:
		error: dereferencing pointer to incomplete type
		current->thread.addr_limit = fs;

It turns out <asm/uaccess.h> was included twice but we don't need
anything from that header.  Same for <linux/stat.h> and a number of
other extraneous headers.

Sort headers alphabetically so it is easier to avoid duplicates.

Test-Parameters: trivial
Change-Id: Ic24a58bdaec16f92d7bea9c24172796031cad471
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>