LU-10560 llite: remove extra headers from rw26.c
Remove headers from llite_mmap.c that are no longer (or were possibly
never) needed.  This avoids a compile problem with 4.14 kernels:

	CC [M]  lustre/llite/rw26.o
	In file included from lustre/llite/rw26.c:43:0:
	./arch/x86/include/asm/uaccess.h: In function ‘set_fs’:
	./arch/x86/include/asm/uaccess.h:31:9:
	error: dereferencing pointer to incomplete type
	current->thread.addr_limit = fs;

It turns out <asm/uaccess.h> was included twice but we don't need
anything from that header.  Same for <linux/stat.h> and a number of
other extraneous headers.

Sort headers alphabetically so it is easier to avoid duplicates.

Change-Id: I4cbe2c1c7289466396b2bb2eac3c475d1041a283
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>